backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=13.0
numpy>=2.0.0
pycryptodome>=3.20.0
python-dotenv>=1.0.0
pydantic>=2.10.0
